+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Version 0.0.21
 
 - Support Angular generate schematics for regular Angular projects
+- Fix to show editor preview properly for the first time
 
 ### Version 0.0.20